Sketched out use of attrs to define / validate class attributes
